checkpoint(library/Byte): moves "Byte/Sequence" to "Sequence/Byte"
