+++ src/library/Byte/exports.ts
@@ -1,4 +1,4 @@
-export * as Sequence from './Sequence';
+export * as Sequence from '../Sequence/Byte';
 
 export {
   Byte_cofactorTo as cofactorTo,
