+++ src/library/Sequence/Byte/EncodingCompatibilityError.ts
