+++ src/library/Sequence/Byte/index.ts
